requirements.txt
streamlit
pandas
Pillow
pyzbar
openpyxl